fix: move patch to prestartup_script.py
this will be executed before the nodes are loaded, importantly even if the nodes fail to load for whatever reason
